GBRForestWriter to single cc file
